Move getAdditionalCertificates function to a dedicated class (no change in the implementation).
